requirements.txt
pandas==2.2.2
yfinance==0.2.38
plotly>=5.18.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
google-api-core>=2.17.0